Extract ListAppEnvVar to its own types module
The legacy CfAppVariablesDataSource hosted the ListAppEnvVar
interface as a side-export — five consumers (signal-config,
variables-tab, table-cell-edit-variable + its spec, app-variables
actions, application-env-var action-builders) all pulled it from the
data-source file. Lifting the type to a tiny cf-app-variables.types.ts
unblocks deleting the now-vestigial data-source + legacy list-config
in the next slice.

Pure import-path shuffle; ListAppEnvVar shape is unchanged.
