--- src/frontend/packages/cloud-foundry/src/shared/components/list/list-types/app-instance/cf-app-instances-config.service.ts
@@ -1,270 +1,0 @@
-import { Injectable, inject } from '@angular/core';
-import { Router } from '@angular/router';
-import { EndpointsDataService, Store } from '@stratosui/store';
-import { combineLatest as combineLatestObs, Observable } from 'rxjs';
-import { combineLatest, map, switchMap } from 'rxjs/operators';
-import {
-  CurrentUserPermissionsService,
-  UtilsService,
-  ConfirmationDialogConfig,
-  ConfirmationDialogService,
-  getIntegerFieldSortFunction,
-  ITableColumn,
-  IGlobalListAction,
-  IListAction,
-  IListConfig,
-  IMultiListAction,
-  ListViewTypes,
-} from '@stratosui/core';
-import {
-  EntityServiceFactory,
-  PaginationMonitorFactory,
-  MetricQueryConfig,
-  IMetricMatrixResult,
-  IMetrics,
-  IMetricApplication,
-  MetricQueryType,
-} from '@stratosui/store';
-import { DeleteApplicationInstance } from '../../../../../actions/application.actions';
-import { FetchApplicationMetricsAction } from '../../../../../actions/cf-metrics.actions';
-import { CFAppState } from '../../../../../cf-app-state';
-import { ApplicationService } from '../../../../../features/applications/application.service';
-import { CfCellHelper } from '../../../../../features/cf/cf-cell.helpers';
-import { CfCurrentUserPermissions } from '../../../../../user-permissions/cf-user-permissions-checkers';
-import { ListAppInstance } from './app-instance-types';
-import { CfAppInstancesDataSource } from './cf-app-instances-data-source';
-import { TableCellCfCellComponent } from './table-cell-cf-cell/table-cell-cf-cell.component';
-import { TableCellUsageComponent } from './table-cell-usage/table-cell-usage.component';
-
-export function createAppInstancesMetricAction(appGuid: string, cfGuid: string): FetchApplicationMetricsAction {
-  return new FetchApplicationMetricsAction(
-    appGuid,
-    cfGuid,
-    new MetricQueryConfig('firehose_container_metric_cpu_percentage'),
-    MetricQueryType.QUERY
-  );
-}
-
-@Injectable({
-  providedIn: 'root'
-})
-export class CfAppInstancesConfigService implements IListConfig<ListAppInstance> {
-  private store = inject<Store<CFAppState>>(Store);
-  private appService = inject(ApplicationService);
-  private utilsService = inject(UtilsService);
-  private router = inject(Router);
-  private confirmDialog = inject(ConfirmationDialogService);
-
-
-  instancesSource: CfAppInstancesDataSource;
-  metricResults$!: Observable<IMetricMatrixResult<IMetricApplication>[]>;
-  columns: Array<ITableColumn<ListAppInstance>> = [
-    {
-      columnId: 'index',
-      headerCell: () => 'Index',
-      cellDefinition: {
-        getValue: (row) => `${row.index}`
-      },
-      sort: getIntegerFieldSortFunction('index'),
-      cellFlex: '1'
-    },
-    {
-      columnId: 'state',
-      headerCell: () => 'State',
-      cellDefinition: {
-        getValue: (row) => `${row.value.state}`
-      },
-      sort: {
-        type: 'sort',
-        orderKey: 'state',
-        field: 'value.state'
-      }, cellFlex: '1',
-      class: 'app-table__cell--table-column-nowrap',
-    },
-    {
-      columnId: 'memory', headerCell: () => 'Memory',
-      cellConfig: {
-        value: (row: ListAppInstance): number => row.usage.mem,
-        label: (row: ListAppInstance): string => this.utilsService.usageBytes([
-          row.usage.hasStats ? row.value.stats.usage.mem : 0,
-          row.usage.hasStats ? row.value.stats.mem_quota : 0
-        ])
-      },
-      cellComponent: TableCellUsageComponent, sort: {
-        type: 'sort',
-        orderKey: 'memory',
-        field: 'usage.mem'
-      }, cellFlex: '2'
-    },
-    {
-      columnId: 'disk', headerCell: () => 'Disk',
-      cellConfig: {
-        value: (row: ListAppInstance): number => row.usage.disk,
-        label: (row: ListAppInstance): string => this.utilsService.usageBytes([
-          row.usage.hasStats ? row.value.stats.usage.disk : 0,
-          row.usage.hasStats ? row.value.stats.disk_quota : 0
-        ])
-      },
-      cellComponent: TableCellUsageComponent, sort: {
-        type: 'sort',
-        orderKey: 'disk',
-        field: 'usage.disk'
-      }, cellFlex: '2'
-    },
-    {
-      columnId: 'cpu', headerCell: () => 'CPU',
-      cellConfig: {
-        value: (row: ListAppInstance): number => row.usage.cpu,
-        label: (row: ListAppInstance): string => this.utilsService.percent(row.usage.hasStats ? row.value.stats.usage.cpu : 0)
-      },
-      cellComponent: TableCellUsageComponent, sort: {
-        type: 'sort',
-        orderKey: 'cpu',
-        field: 'usage.cpu'
-      }, cellFlex: '2'
-    },
-    {
-      columnId: 'uptime',
-      headerCell: () => 'Uptime',
-      cellDefinition: {
-        getValue: (row) => row.usage.hasStats ? this.utilsService.formatUptime(row.value.stats.uptime) : '-'
-      },
-      sort: {
-        type: 'sort',
-        orderKey: 'uptime',
-        field: 'value.stats.uptime'
-      }, cellFlex: '3'
-    }
-  ];
-  cfCellColumn: ITableColumn<ListAppInstance> = {
-    columnId: 'cell',
-    headerCell: () => 'Cell',
-    cellConfig: {
-      metricResults$: null
-    },
-    cellComponent: TableCellCfCellComponent,
-    cellFlex: '2'
-  };
-
-  viewType = ListViewTypes.TABLE_ONLY;
-  enableTextFilter = true;
-  text: { title: string | null; filter: string; noEntries: string } = {
-    title: null,
-    filter: 'Filter by State',
-    noEntries: 'There are no application instances'
-  };
-  private initialised$: Observable<boolean>;
-
-  private listActionTerminate: IListAction<any> = {
-    action: (item: any) => {
-      const confirmation = new ConfirmationDialogConfig(
-        'Terminate Instance?',
-        `Are you sure you want to terminate instance ${item.index}?`,
-        'Terminate',
-        true
-      );
-      this.confirmDialog.open(
-        confirmation,
-        () => this.store.dispatch(new DeleteApplicationInstance(this.appService.appGuid, item.index, this.appService.cfGuid))
-      );
-    },
-    label: 'Terminate',
-    description: ``, // Description depends on console user permission
-    createVisible: () => this.canEditApp$
-  };
-
-  private listActionSsh: IListAction<any> = {
-    action: (item: any) => {
-      const index = item.index;
-      const sshRoute = (
-        `/applications/${this.appService.cfGuid}/${this.appService.appGuid}/ssh/${index}`
-      );
-      this.router.navigate([sshRoute]);
-    },
-    label: 'SSH',
-    description: ``, // Description depends on console user permission
-    createEnabled: row$ =>
-      row$.pipe(switchMap(row => {
-        return this.appService.app$.pipe(
-          combineLatest(this.appService.appSpace$),
-          map(([app, space]) => {
-            return !!(
-              row &&
-              row.value &&
-              row.value.state === 'RUNNING' &&
-              app &&
-              app.entity &&
-              app.entity.entity &&
-              app.entity.entity.enable_ssh &&
-              space &&
-              space.allowSsh
-            );
-          })
-        );
-      })),
-    createVisible: () => this.canEditApp$
-  };
-
-  private singleActions = [
-    this.listActionTerminate,
-    this.listActionSsh,
-  ];
-
-  private canEditApp$: Observable<boolean>;
-
-  constructor() {
-    const store = this.store;
-    const appService = this.appService;
-    const entityServiceFactory = inject(EntityServiceFactory);
-    const paginationMonitorFactory = inject(PaginationMonitorFactory);
-    const cups = inject(CurrentUserPermissionsService);
-
-    const endpointsService = inject(EndpointsDataService);
-    const cellHelper = new CfCellHelper(store, paginationMonitorFactory, endpointsService);
-
-    this.initialised$ = cellHelper.hasCellMetrics(appService.cfGuid).pipe(
-      map(hasMetrics => {
-        if (hasMetrics) {
-          this.columns.splice(1, 0, this.cfCellColumn);
-          this.cfCellColumn.cellConfig = {
-            metricEntityService: this.createMetricsResults(entityServiceFactory),
-            cfGuid: this.appService.cfGuid
-          };
-        }
-        return true;
-      })
-    );
-
-    this.instancesSource = new CfAppInstancesDataSource(
-      this.store,
-      this.appService.cfGuid,
-      this.appService.appGuid,
-      this,
-    );
-
-    this.canEditApp$ = combineLatestObs(
-      appService.appOrg$,
-      appService.appSpace$
-    ).pipe(
-      switchMap(([org, space]) =>
-        cups.can(CfCurrentUserPermissions.APPLICATION_EDIT, appService.cfGuid, org!.guid, space!.guid)
-      )
-    );
-  }
-
-  getGlobalActions = (): IGlobalListAction<ListAppInstance>[] => [];
-  getMultiActions = (): IMultiListAction<ListAppInstance>[] => [];
-  getSingleActions = (): IListAction<ListAppInstance>[] => this.singleActions;
-  getColumns = (): ITableColumn<ListAppInstance>[] => this.columns;
-  getDataSource = () => this.instancesSource;
-  getMultiFiltersConfigs = (): any[] => [];
-  getInitialised = (): Observable<boolean> => this.initialised$;
-
-  private createMetricsResults(entityServiceFactory: EntityServiceFactory) {
-    const metricsAction = createAppInstancesMetricAction(this.appService.appGuid, this.appService.cfGuid);
-    return entityServiceFactory.create<IMetrics<IMetricMatrixResult<IMetricApplication>>>(
-      metricsAction.guid,
-      metricsAction
-    );
-  }
-}
