+++ src/frontend/packages/cloud-foundry/src/shared/components/list/list-types/app-variables/cf-app-variables.types.ts
@@ -1,0 +1,4 @@
+export interface ListAppEnvVar {
+  name: string;
+  value: string;
+}
