+++ src/frontend/packages/cloud-foundry/src/shared/components/list/list-types/app-variables/table-cell-edit-variable/table-cell-edit-variable.component.spec.ts
@@ -1,4 +1,4 @@
-import { ListAppEnvVar } from '../cf-app-variables-data-source';
+import { ListAppEnvVar } from '../cf-app-variables.types';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { provideZonelessChangeDetection } from '@angular/core';
 import { describe, it, expect, beforeEach } from 'vitest';
